# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quality Incident Reporting System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  supervisor: "supervisor",
  operator: "operator",
} as const;

export type Department = (typeof Department)[keyof typeof Department];

export const Department = {
  Production: "Production",
  Quality: "Quality",
  Engineering: "Engineering",
  Warehouse: "Warehouse",
  HSE: "HSE",
} as const;

export type IncidentCategory =
  (typeof IncidentCategory)[keyof typeof IncidentCategory];

export const IncidentCategory = {
  Contamination: "Contamination",
  Underweight: "Underweight",
  Equipment_Failure: "Equipment Failure",
  Process_Deviation: "Process Deviation",
  Packaging_Defect: "Packaging Defect",
  Others: "Others",
} as const;

export type Severity = (typeof Severity)[keyof typeof Severity];

export const Severity = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
  Critical: "Critical",
} as const;

export type IncidentStatus =
  (typeof IncidentStatus)[keyof typeof IncidentStatus];

export const IncidentStatus = {
  Open: "Open",
  In_Progress: "In Progress",
  Closed: "Closed",
} as const;

export interface User {
  id: number;
  email: string;
  fullName: string;
  role: UserRole;
  createdAt: string;
}

export interface RegisterBody {
  /** @minLength 3 */
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 1 */
  fullName: string;
  role?: UserRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
}

export interface Lookups {
  departments: Department[];
  categories: IncidentCategory[];
  severities: Severity[];
  statuses: IncidentStatus[];
  lines: string[];
  productTypes: string[];
  roles: UserRole[];
}

export interface Incident {
  id: number;
  incidentCode: string;
  occurredAt: string;
  department: Department;
  line: string;
  productType: string;
  category: IncidentCategory;
  severity: Severity;
  description: string;
  immediateAction: string;
  status: IncidentStatus;
  attachmentUrl?: string | null;
  rootCauseCategory?: string | null;
  reportedById: number;
  reportedByName: string;
  assignedToId?: number | null;
  assignedToName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateIncidentBody {
  department: Department;
  /** @minLength 1 */
  line: string;
  /** @minLength 1 */
  productType: string;
  category: IncidentCategory;
  severity: Severity;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  immediateAction: string;
  attachmentUrl?: string | null;
  rootCauseCategory?: string | null;
}

export interface UpdateIncidentBody {
  status?: IncidentStatus;
  immediateAction?: string;
  rootCauseCategory?: string | null;
  assignedToId?: number | null;
}

export interface UserSummary {
  id: number;
  fullName: string;
  role: UserRole;
}

export interface CategoryCount {
  category: IncidentCategory;
  count: number;
}

export interface DepartmentCount {
  department: Department;
  count: number;
}

export interface SeverityCount {
  severity: Severity;
  count: number;
}

export interface StatusCount {
  status: IncidentStatus;
  count: number;
}

export interface TrendPoint {
  date: string;
  count: number;
}

export type DashboardSummaryTrendByDepartmentItem = {
  date: string;
  [key: string]: unknown;
};

export interface DashboardSummary {
  totalIncidents: number;
  openIncidents: number;
  criticalIncidents: number;
  last30DaysIncidents: number;
  byCategory: CategoryCount[];
  byDepartment: DepartmentCount[];
  bySeverity: SeverityCount[];
  byStatus: StatusCount[];
  trend: TrendPoint[];
  trendByDepartment: DashboardSummaryTrendByDepartmentItem[];
  departments: string[];
  recent: Incident[];
}

export type ListIncidentsParams = {
  search?: string;
  department?: Department;
  category?: IncidentCategory;
  severity?: Severity;
  status?: IncidentStatus;
  startDate?: string;
  endDate?: string;
  reportedById?: number;
  assignedToId?: number;
};