# lib/api-zod/src/generated/types/incident.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quality Incident Reporting System API
 * OpenAPI spec version: 0.1.0
 */
import type { Department } from "./department";
import type { IncidentCategory } from "./incidentCategory";
import type { IncidentStatus } from "./incidentStatus";
import type { Severity } from "./severity";

export interface Incident {
  id: number;
  incidentCode: string;
  occurredAt: Date;
  department: Department;
  line: string;
  productType: string;
  category: IncidentCategory;
  severity: Severity;
  description: string;
  immediateAction: string;
  status: IncidentStatus;
  attachmentUrl?: string | null;
  rootCauseCategory?: string | null;
  reportedById: number;
  reportedByName: string;
  assignedToId?: number | null;
  assignedToName?: string | null;
  createdAt: Date;
  updatedAt: Date;
}